requirements.txt
attrs==22.1.0
Automat==22.10.0
autopep8==2.0.1
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
cymem==2.0.7
dnspython==2.2.1
filelock==3.8.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.9
numpy==1.24.0
packaging==22.0
pandas==1.5.2
parsel==1.7.0
pathy==0.10.1
preshed==3.0.8
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.6
pymongo==4.3.3
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
srsly==2.4.5
thinc==8.1.6
tldextract==3.4.0
tomli==2.0.1
tqdm==4.64.1
Twisted==22.10.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
wasabi==0.10.1
wget==3.2
zope.interface==5.5.2
